Refactor #358 - make home module
